be_hosp_enCasa/requirements.txt
asgiref==3.5.2
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
psycopg2-binary==2.9.3
PyJWT==2.5.0
pytz==2022.4
sqlparse==0.4.3
gunicorn==20.1.0
django-heroku==0.3.1